3 datasets potentially to use in project
In every workbook there is a link to source and transformations made. The difficulty is that number of features has to be at least 50% the number of observations. In parkinson and swarm datasets that is obtained through transformations. In wine dataset permutations of present columns are generated to add features.